Add files via upload
Used to to control a single servo and a 3 aspect signal. More IO are available on the ESP8266 to do other things.
Program set at 160MHz for ESP8266 12E but may vary depending on board. 80MHz would be fine for this project.
Make sure to download the required boards and libraries through the Arduino IDE